# api/openapi/quiver.v1.yaml
basePath: /
definitions:
  internal_api.APIError:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  internal_api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_api.APIError'
      request_id:
        type: string
    type: object
...
  internal_api.ProtocolResponse:
    properties:
      flow_count:
        type: integer
      metric:
        type: string
      protocol_number:
        type: integer
      transport_protocol:
        type: string
      value:
        type: integer
    type: object
  internal_api.ProtocolsResponse:
    properties:
      from:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_api.ProtocolResponse'
        type: array
      limit:
        type: integer
      to:
        type: string
    type: object
  internal_api.TopPortResponse:
    properties:
      flow_count:
        type: integer
      metric:
        type: string
      port:
        type: integer
      value:
        type: integer
    type: object
  internal_api.TopPortsResponse:
    properties:
      from:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_api.TopPortResponse'
        type: array
      limit:
        type: integer
      to:
        type: string
    type: object
  internal_api.TopTalkerResponse:
    properties:
      flow_count:
        type: integer
      ip:
        type: string
      metric:
        type: string
      value:
        type: integer
    type: object
  internal_api.TopTalkersResponse:
    properties:
      from:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_api.TopTalkerResponse'
        type: array
      limit:
        type: integer
      to:
        type: string
    type: object
info:
  contact: {}
  description: 'Network flow ingestion and query API. Protected endpoints use X-API-Key
    scopes: ingest, query, metrics. Generated output is Swagger 2.0.'
  title: Quiver API
  version: "0.1"
paths:
  /api/v1/aggregations/protocols:
    get:
      description: Aggregates flows by protocol number and transport protocol.
      parameters:
      - description: Inclusive event_start_time lower bound
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive event_start_time upper bound
        in: query
        name: to
        required: true
        type: string
      - description: Aggregation metric
        enum:
        - bytes
        - packets
        - flows
        in: query
        name: metric
        type: string
      - description: Maximum number of rows
        in: query
        name: limit
        type: integer
      - description: Source type filter
        enum:
        - netflow_v5
        - netflow_v9
        - zeek_conn_json
        - suricata_eve_json
        - rest_json
        - syslog_cef
        - syslog_leef
        in: query
        name: source_type
        type: string
      - description: Exact source IP filter
        in: query
        name: src_ip
        type: string
      - description: Exact destination IP filter
        in: query
        name: dst_ip
        type: string
      - description: IP protocol number filter
        in: query
        name: protocol_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.ProtocolsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Protocol distribution
      tags:
      - aggregations
  /api/v1/aggregations/top-ports:
    get:
      description: Aggregates flows by source or destination port.
      parameters:
      - description: Inclusive event_start_time lower bound
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive event_start_time upper bound
        in: query
        name: to
        required: true
        type: string
      - description: src or dst
        in: query
        name: direction
        required: true
        type: string
      - description: Aggregation metric
        enum:
        - bytes
        - packets
        - flows
        in: query
        name: metric
        type: string
      - description: Maximum number of rows
        in: query
        name: limit
        type: integer
      - description: Source type filter
        enum:
        - netflow_v5
        - netflow_v9
        - zeek_conn_json
        - suricata_eve_json
        - rest_json
        - syslog_cef
        - syslog_leef
        in: query
        name: source_type
        type: string
      - description: Exact source IP filter
        in: query
        name: src_ip
        type: string
      - description: Exact destination IP filter
        in: query
        name: dst_ip
        type: string
      - description: IP protocol number filter
        in: query
        name: protocol_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.TopPortsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Top ports
      tags:
      - aggregations
  /api/v1/aggregations/top-talkers:
    get:
      description: Aggregates flows by source or destination IP.
      parameters:
      - description: Inclusive event_start_time lower bound
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive event_start_time upper bound
        in: query
        name: to
        required: true
        type: string
      - description: src or dst
        in: query
        name: direction
        required: true
        type: string
      - description: Aggregation metric
        enum:
        - bytes
        - packets
        - flows
        in: query
        name: metric
        type: string
      - description: Maximum number of rows
        in: query
        name: limit
        type: integer
      - description: Source type filter
        enum:
        - netflow_v5
        - netflow_v9
        - zeek_conn_json
        - suricata_eve_json
        - rest_json
        - syslog_cef
        - syslog_leef
        in: query
        name: source_type
        type: string
      - description: Exact source IP filter
        in: query
        name: src_ip
        type: string
      - description: Exact destination IP filter
        in: query
        name: dst_ip
        type: string
      - description: IP protocol number filter
        in: query
        name: protocol_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.TopTalkersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Top talkers
      tags:
      - aggregations
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key with endpoint-specific scope. Ingest endpoints require ingest,
      query and aggregation endpoints require query, and metrics endpoints require
      metrics.
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"